nba_backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.1.1
